Update CI builds (#74)
Our Windows and Mac images were deprecated or directly removed. Ubuntu 18.04 is technically still alive (its a LTS, so it should be on the 5 year cycle), but 22.04 just came out, so it also makes sense to jump to 20.04, which is already 2 years old (especially if we consider the GLIBC issue in memcpy).

These are, basically, the same images RDKit is using now. I'll open another PR for coordgenlibs.

bumps ubuntu and VS target version